Update fetch roles config module
