+++ common/modules/library/fetch_roles_config.py
@@ -17,20 +17,20 @@
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.discovery.omniadb_connection import get_data_from_db
 
-FIRST_LAYER_ROLES = {"service", "login", "compiler", "kube_control_plane", "etcd", "slurm_control_node", "slurm_dbd", "auth_server"}
+MANAGEMENT_LAYER_ROLES = {"service", "login", "compiler", "kube_control_plane", "etcd", "slurm_control_node", "slurm_dbd", "auth_server"}
 SECOND_LAYER_ROLES = {"default", "kube_node", "slurm_node"}
-NON_SERVICE_ROLES = (FIRST_LAYER_ROLES | SECOND_LAYER_ROLES) - {"service"}
+NON_SERVICE_ROLES = (MANAGEMENT_LAYER_ROLES | SECOND_LAYER_ROLES) - {"service"}
 
-def validate_roles(roles, layer, module, first_layer_roles=FIRST_LAYER_ROLES, second_layer_roles=SECOND_LAYER_ROLES, non_service_roles=NON_SERVICE_ROLES):
+def validate_roles(roles, layer, module, management_layer_roles=MANAGEMENT_LAYER_ROLES, second_layer_roles=SECOND_LAYER_ROLES, non_service_roles=NON_SERVICE_ROLES):
     """
     Validates roles based on multiple conditions:
-    1. Roles should only belong to either first-layer or compute-layer roles.
+    1. Roles should only belong to either management_layer or compute-layer roles.
     2. At least one role should exist in the given layer.
-    3. Groups associated with first-layer roles should not be in compute-layer roles.
-    4. Groups assigned to 'service' should not be in other first-layer roles.
+    3. Groups associated with management_layer roles should not be in compute-layer roles.
+    4. Groups assigned to 'service' should not be in other management_layer roles.
 
     :param roles: Dictionary where keys are role names and values are dictionaries with a 'groups' key containing a list of groups.
-    :param first_layer_roles: Set of first-layer role names.
+    :param management_layer_roles: Set of management_layer role names.
     :param compute_layer_roles: Set of compute-layer role names.
     :param layer: Specifies which layer should have at least one role. Should be 'first' or 'default'.
     :raises RoleValidationError: If validation fails, raises an exception with the list of errors.
@@ -42,16 +42,16 @@ def validate_roles(roles, layer, module, first_layer_roles=FIRST_LAYER_ROLES, se
 
     defined_roles = set(roles.keys())  # Extract all roles from input
 
-    # Check 1: Ensure all roles belong to either first-layer or compute-layer roles
-    invalid_roles = defined_roles - (first_layer_roles | second_layer_roles)
+    # Check 1: Ensure all roles belong to either management_layer or compute-layer roles
+    invalid_roles = defined_roles - (management_layer_roles | second_layer_roles)
     errors = []
     if invalid_roles:
-        module.warn(f"Invalid roles detected: {invalid_roles}. Roles must be from either first-layer or compute-layer roles.")
+        module.warn(f"Invalid roles detected: {invalid_roles}. Roles must be from either management_layer or compute-layer roles.")
 
     # Check 1&2: Ensure at least one role exists in the specified layer
     if layer == "first":
-        if not defined_roles.intersection(first_layer_roles):
-            raise Exception("At least one role must be from the first-layer roles.")
+        if not defined_roles.intersection(management_layer_roles):
+            raise Exception("At least one role must be from the management_layer roles.")
     else:
         if 'service' in defined_roles:
             if not defined_roles.intersection(second_layer_roles):
@@ -60,23 +60,23 @@ def validate_roles(roles, layer, module, first_layer_roles=FIRST_LAYER_ROLES, se
             if not defined_roles.intersection(non_service_roles):
                 raise Exception(f"At least one role must be defined from - {non_service_roles} roles_config.yml")
 
-    # Collect all groups used by first-layer and compute-layer roles
-    first_layer_groups = {group for role in first_layer_roles for group in role_groups.get(role, [])}
+    # Collect all groups used by management_layer and compute-layer roles
+    management_layer_groups = {group for role in management_layer_roles for group in role_groups.get(role, [])}
     second_layer_groups = {group for role in second_layer_roles for group in role_groups.get(role, [])}
 
-    # Check 3: Ensure groups from first-layer roles are not in compute-layer roles
-    common_groups = first_layer_groups.intersection(second_layer_groups)
+    # Check 3: Ensure groups from management_layer roles are not in compute-layer roles
+    common_groups = management_layer_groups.intersection(second_layer_groups)
     if common_groups:
-        errors.append(f"Groups {common_groups} are assigned to both first-layer and compute-layer roles.")
+        errors.append(f"Groups {common_groups} are assigned to both management_layer and compute-layer roles.")
 
-    # Check 4: Ensure groups in 'service' role are not part of other first-layer roles
+    # Check 4: Ensure groups in 'service' role are not part of other management_layer roles
     service_groups = role_groups.get("service", set())
 
-    for role in first_layer_roles:
+    for role in management_layer_roles:
         if role != "service":
             overlapping_groups = service_groups.intersection(role_groups.get(role, set()))
             if overlapping_groups:
-                errors.append(f"Groups {overlapping_groups} from 'service' role are also part of first-layer role '{role}'.")
+                errors.append(f"Groups {overlapping_groups} from 'service' role are also part of management_layer role '{role}'.")
 
     # Raise an error if validation fails
     if errors:
@@ -129,7 +129,7 @@ def filter_roles(groups_data, roles_data, layer):
     """Filter the roles based on the layer and the roles data."""
 
     if layer == "first":
-        valid_roles = set(roles_data.keys()).intersection(FIRST_LAYER_ROLES)
+        valid_roles = set(roles_data.keys()).intersection(MANAGEMENT_LAYER_ROLES)
     else:
         if 'service' in roles_data:
             valid_roles = set(roles_data.keys()).intersection(SECOND_LAYER_ROLES)
@@ -151,12 +151,15 @@ def roles_groups_mapping(groups_data, roles_data, layer):
         tuple: A tuple containing the following:
             - bmc_check (bool): A boolean indicating if BMC is required.
             - switch_check (bool): A boolean indicating if switch is required.
+            - hierarchical_provision_status (bool): A boolean indicating if hierarchical provisioning is required.
             - roles_groups_data (dict): A dictionary containing the roles and groups data.
             - groups_roles_info (dict): A dictionary containing the groups and roles information.
+            - hierarchical_service_data (dict): A dictionary containing the hierarchical service node data.
 
     Raises:
         Exception: If a group doesn't exist in the roles_config.yml Groups dict.
     """
+
 
     valid_roles = filter_roles(groups_data, roles_data, layer)
 
@@ -174,9 +177,11 @@ def roles_groups_mapping(groups_data, roles_data, layer):
                 groups_roles_info.setdefault(group, {}).setdefault('roles', []).append(role)
                 groups_roles_info[group].update(groups_data.get(group))
                 grp_bmc_check = check_bmc_required(groups_data[group])
-                bmc_check = bmc_check or grp_bmc_check
                 grp_switch_check = grp_bmc_check and check_switch_required(groups_data[group], layer)
+                # For a group bmc will be false if switch is true
+                grp_bmc_check = False if grp_switch_check else grp_bmc_check
                 switch_check = switch_check or grp_switch_check
+                bmc_check = bmc_check or grp_bmc_check
                 service_node = check_hierarchical_provision(group, groups_data, roles_data, layer)
                 if service_node:
                     snode_name = list(service_node.keys())[0]
