+++ local_repo/ansible.cfg
@@ -7,7 +7,8 @@ timeout = 180
 executable = /bin/bash
 stdout_callback = default
 callbacks_enabled = profile_tasks
-library = ../input_validation/library:roles/parse_and_download/library
+library = ../input_validation/library:roles/parse_and_download/library:/omnia/common/modules/library
+module_utils = /omnia/common/modules/module_utils
 
 [persistent_connection]
 command_timeout = 180
